aider: Add command to compile help tags after all plugins have been loaded.
# Aider chat conversation:

USER: What configuration settings for the vim-polyglot plugin is available? I don't mean settings for the bundled plugins, but the polyglot plugin itself.

ASSISTANT: The `vim-polyglot` plugin is a collection of language packs for Vim that provides syntax highlighting, indentation, and other language-specific features. It is designed to be a comprehensive bundle that is easy to install and use. However, `vim-polyglot` itself has minimal configuration options because it acts as an aggregator of other plugins.

The primary settings available for `vim-polyglot` are related to disabling specific language packs or features that you do not want to use, or that might conflict with other plugins you have installed. Here are some of the settings you can use in your `.vimrc` or equivalent Vim configuration file:

1. `g:polyglot_disabled`: This variable is a list of language packs that you want to disable. For example, to disable the language packs for `ruby` and `python`, you would set it like this:
   ```vim
   let g:polyglot_disabled = ['ruby', 'python']
   ```

2. `g:polyglot_disabled_groups`: This variable allows you to disable groups of related language packs. For instance, to disable all auto-indentation features provided by `vim-polyglot`, you would set it like this:
   ```vim
   let g:polyglot_disabled_groups = ['autoindent']
   ```

3. `g:polyglot_is_disabled`: This function can be used to check if a specific language pack is disabled. It can be useful in conditional configurations.

Remember that `vim-polyglot` is designed to work out of the box with sensible defaults for each language, so typically, you would not need to configure much beyond disabling certain language packs if they are not needed or if you prefer to use another plugin for that language.

For the most up-to-date and detailed information on configuring `vim-polyglot`, you should refer to the official [vim-polyglot repository](https://github.com/sheerun/vim-polyglot) on GitHub.

USER: does g:ale_disable_lsp still need to be set before enabling filetype settings?

ASSISTANT: The `g:ale_disable_lsp` variable in ALE (Asynchronous Lint Engine) is used to disable ALE's Language Server Protocol (LSP) features. This might be necessary if you are using another LSP client, like CoC (Conquer of Completion), and you want to avoid conflicts or duplicated functionality.

As of my knowledge cutoff in early 2023, you would typically set `g:ale_disable_lsp` in your `.vimrc` or equivalent Vim configuration file before ALE is loaded to ensure that ALE's LSP features are disabled. This is usually done globally and not on a per-filetype basis.

Here is an example of how you might set it in your `.vimrc`:

```vim
let g:ale_disable_lsp = 1
```

However, ALE's handling of LSP features and the need for this setting could change over time. It's always a good practice to check the latest documentation or the GitHub repository of ALE for the most current information regarding configuration settings.

Since the `.vimrc` and ALE plugin configuration files are marked as *read-only* in the provided file list, I cannot propose changes to these files. If you would like to make changes to any of these files, please add them to the chat as *read-write* files.

USER: help is not being compiled correctly. If I put a 'helptags ALL' command in an after section, would that resolve the issue?

ASSISTANT: To ensure that help tags are compiled correctly, you can indeed use the `:helptags ALL` command. This command should be executed after all help files have been updated or added. Placing this command in an "after" section, such as in an `after/plugin` script, would cause it to run after Vim has loaded all the plugins, which should include the help files.

However, since the `.vimrc` and other Vim configuration files are marked as *read-only* in the provided file list, I cannot create a diff for these files. If you want to add this command to your Vim configuration, you would typically add it to a Vimscript file that is sourced after all plugins have been loaded.

If you have a specific file in mind where you want to add the `:helptags ALL` command, and it is not listed as *read-only*, please add it to the chat as a *read-write* file, and I can then provide you with the necessary diff to make the change. If you need to create a new Vimscript file for this purpose, please let me know, and I can show you the diff for creating a new file with this command.

USER: I added these *read-write* files: .vimrc

ASSISTANT: To ensure that help tags are compiled correctly, I will add a command to the `.vimrc` file to generate help tags after all plugins have been loaded. This will be
^C KeyboardInterrupt

USER: There is already a helptags all in the .vimrc. I think it needs to be in .vim/pack/settings/after ... but I'm not sure past that ... .vim/pack/settings/start/settings/after/plugin/mustbelast.vim ?

ASSISTANT: To ensure that the `:helptags` command is executed after all plugins have been loaded, we can create a new Vimscript file in the `.vim/pack/settings/start/settings/after/plugin/` directory. The file `mustbelast.vim` is a good choice for a name, as it implies that this script should run last.

Here is the diff for creating a new file named `.vim/pack/settings/start/settings/after/plugin/mustbelast.vim` with the `:helptags ALL` command:

```diff
--- /dev/null
+++ .vim/pack/settings/start/settings/after/plugin/mustbelast.vim
@@ ... @@
+" Compile help tags after all plugins have been loaded
+silent! helptags ALL
```

This will create a new file that, when sourced, will run the `:helptags ALL` command silently, compiling all help tags for the help documentation provided by the plugins. The use of `silent!` ensures that any errors (such as missing help files) do not interrupt the user's workflow.